Opcode fix.

+++ Trunk/src/hearthstone-world/Player.cpp
@@ -6084,7 +6084,7 @@ void Player::AddInRangeObject(Object* pObj)
 
 					data << uint8(aur->m_auraSlot);
 					data << uint32(aur->GetSpellId());
-					data << uint16(flags);
+					data << uint8(flags);
 					data << uint8(aur->GetUnitCaster() ? aur->GetUnitCaster()->getLevel() : 0);
 					data << uint8(aur->stackSize > 1 ? aur->stackSize : aur->procCharges ? aur->procCharges : 1);
